Add namespace to scripts
