+++ Assets/Scripts/Character/BaseCharacter.cs
@@ -1,41 +1,44 @@
 using UnityEngine;
 
-public class BaseCharacter : MonoBehaviour
+namespace GGJ.Characters
 {
-    //[Header("Setting"), SerializeField] LayerMask interactiveLayer;
-
-    protected virtual void Awake()
+    public class BaseCharacter : MonoBehaviour
     {
-        Initialize();
-    }
+        //[Header("Setting"), SerializeField] LayerMask interactiveLayer;
 
-    protected virtual void Update()
-    {
-        PlayAnimation();
-    }
+        protected virtual void Awake()
+        {
+            Initialize();
+        }
 
-    protected virtual void FixedUpdate()
-    {
+        protected virtual void Update()
+        {
+            PlayAnimation();
+        }
 
-    }
+        protected virtual void FixedUpdate()
+        {
 
-    protected virtual void Initialize()
-    {
+        }
 
-    }
+        protected virtual void Initialize()
+        {
 
-    protected virtual void PlayAnimation()
-    {
+        }
 
-    }
+        protected virtual void PlayAnimation()
+        {
 
-    protected virtual void OnTriggerEnter2D(Collider2D collision)
-    {
-        
-    }
+        }
 
-    protected virtual void OnTriggerExit2D(Collider2D collision)
-    {
+        protected virtual void OnTriggerEnter2D(Collider2D collision)
+        {
 
+        }
+
+        protected virtual void OnTriggerExit2D(Collider2D collision)
+        {
+
+        }
     }
 }
